Clean up project
There is still a lot to do for this project, but the main functionality is apt. The development will not cease, but will be offloaded to other branches for future merges.

Specific changes include
- Readme.md
Turned the Readme into documentation.

- arduinocom.js
Removed extra logPorts() and other developmental code.

- comwindow.html
Not needed anymore. Artifact of previous alogrithm.

- main.js
Removed unneeded references to Comwindow,
Removed unnecessary lines of code (slimming down)

- package-lock.json
Honestly probably the auto-formatter

- package.json
Changed name of package
Added scripts to build program for OSX and Windows and Linux
Changed version number to 0.8.0
Included electron-packager